Rename notebook text diff editor (#164149)
Rename notebook text diff editor.
