+++ src/vs/workbench/contrib/notebook/browser/diff/notebookDiffList.ts
